file tree re-structure
